Moved all github.com instrumentation plugins below github.com
Fix up all example docker-compose/Dockerfile
