--- instrumentation/gin-gonic/gin/example/docker-compose.yml
@@ -1,39 +1,0 @@
-# Copyright The OpenTelemetry Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-version: "3.7"
-services:
-  gin-client:
-    image: golang:alpine
-    networks:
-      - example
-    command:
-      - "/bin/sh"
-      - "-c"
-      - "wget http://gin-server:8080/users/123 && cat 123"
-    depends_on:
-      - gin-server
-  gin-server:
-    build:
-      dockerfile: $PWD/Dockerfile
-      context: ../../../..
-    ports:
-      - "8080:80"
-    command:
-      - "/bin/sh"
-      - "-c"
-      - "/go/bin/server"
-    networks:
-      - example
-networks:
-  example:
