+++ instrumentation/github.com/labstack/echo/example/server.go
@@ -1,0 +1,78 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"context"
+	"log"
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+
+	echotrace "go.opentelemetry.io/contrib/instrumentation/github.com/labstack/echo"
+	otelglobal "go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/api/kv"
+	oteltrace "go.opentelemetry.io/otel/api/trace"
+	oteltracestdout "go.opentelemetry.io/otel/exporters/trace/stdout"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+)
+
+var tracer = otelglobal.Tracer("gin-server")
+
+func main() {
+	initTracer()
+	r := echo.New()
+	r.Use(echotrace.Middleware("my-server"))
+
+	r.GET("/users/:id", func(c echo.Context) error {
+		id := c.Param("id")
+		name := getUser(c.Request().Context(), id)
+		return c.JSON(http.StatusOK, struct {
+			ID   string
+			Name string
+		}{
+			ID:   id,
+			Name: name,
+		})
+	})
+	_ = r.Start(":8080")
+}
+
+func initTracer() {
+	exporter, err := oteltracestdout.NewExporter(oteltracestdout.Options{PrettyPrint: true})
+	if err != nil {
+		log.Fatal(err)
+	}
+	cfg := sdktrace.Config{
+		DefaultSampler: sdktrace.AlwaysSample(),
+	}
+	tp, err := sdktrace.NewProvider(
+		sdktrace.WithConfig(cfg),
+		sdktrace.WithSyncer(exporter),
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
+	otelglobal.SetTraceProvider(tp)
+}
+
+func getUser(ctx context.Context, id string) string {
+	_, span := tracer.Start(ctx, "getUser", oteltrace.WithAttributes(kv.String("id", id)))
+	defer span.End()
+	if id == "123" {
+		return "echotrace tester"
+	}
+	return "unknown"
+}
