--- instrumentation/gorilla/mux/example/server.go
@@ -1,75 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"net/http"
-
-	"github.com/gorilla/mux"
-
-	muxtrace "go.opentelemetry.io/contrib/instrumentation/gorilla/mux"
-	otelglobal "go.opentelemetry.io/otel/api/global"
-	otelkv "go.opentelemetry.io/otel/api/kv"
-	oteltrace "go.opentelemetry.io/otel/api/trace"
-	oteltracestdout "go.opentelemetry.io/otel/exporters/trace/stdout"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-)
-
-var tracer = otelglobal.Tracer("mux-server")
-
-func main() {
-	initTracer()
-	r := mux.NewRouter()
-	r.Use(muxtrace.Middleware("my-server"))
-	r.HandleFunc("/users/{id:[0-9]+}", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		id := vars["id"]
-		name := getUser(r.Context(), id)
-		reply := fmt.Sprintf("user %s (id %s)\n", name, id)
-		_, _ = w.Write(([]byte)(reply))
-	}))
-	http.Handle("/", r)
-	_ = http.ListenAndServe(":8080", nil)
-}
-
-func initTracer() {
-	exporter, err := oteltracestdout.NewExporter(oteltracestdout.Options{PrettyPrint: true})
-	if err != nil {
-		log.Fatal(err)
-	}
-	cfg := sdktrace.Config{
-		DefaultSampler: sdktrace.AlwaysSample(),
-	}
-	tp, err := sdktrace.NewProvider(
-		sdktrace.WithConfig(cfg),
-		sdktrace.WithSyncer(exporter),
-	)
-	if err != nil {
-		log.Fatal(err)
-	}
-	otelglobal.SetTraceProvider(tp)
-}
-
-func getUser(ctx context.Context, id string) string {
-	_, span := tracer.Start(ctx, "getUser", oteltrace.WithAttributes(otelkv.String("id", id)))
-	defer span.End()
-	if id == "123" {
-		return "muxtrace tester"
-	}
-	return "unknown"
-}
